Try update for performance

+++ src/AWLQtDemo/AWLCoreClasses/DetectionStruct.cpp
@@ -183,7 +183,7 @@ void AScan::FindMinMaxMean(float *min, float *max, float *mean)
 		*max = 0;
 	}
 
-	for (i = sampleOffset; i < sampleCount ; i ++) {
+	for (i = (int)sampleOffset; i < sampleCount ; i ++) {
 		switch (sampleSize) {
 		default:
 				return;
@@ -200,12 +200,12 @@ void AScan::FindMinMaxMean(float *min, float *max, float *mean)
 			break;
 		case 4:
 			if (sampleSigned) {
-				if (i32[i] < *min) *min = i32[i];
-				if (i32[i] > *max) *max = i32[i];
+				if ((float)i32[i] < *min) *min = (float)i32[i];
+				if ((float)i32[i] > *max) *max = (float)i32[i];
 				*mean += i32[i];
 			} else {
-				if (u32[i] < *min) *min = u32[i];
-				if (u32[i] > *max) *max = u32[i];
+				if ((float)u32[i] < *min) *min = (float)u32[i];
+				if ((float)u32[i] > *max) *max = (float)u32[i];
 				*mean += u32[i];
 			}
 		}
